update svelte packages
