+++ package.json
@@ -12,14 +12,14 @@
 		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "next",
-		"@sveltejs/kit": "next",
-		"@zerodevx/svelte-toast": "^0.7.2",
-		"svelte": "^3.44.0",
-		"svelte-check": "^2.7.1",
-		"svelte-preprocess": "^4.10.6",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.0.0",
+		"@zerodevx/svelte-toast": "^0.9.5",
+		"svelte": "^5.0.0",
+		"svelte-check": "^3.0.0",
+		"svelte-preprocess": "^6.0.0",
 		"tslib": "^2.3.1",
-		"typescript": "^4.7.2"
+		"typescript": "^5.3.3"
 	},
 	"type": "module",
 	"dependencies": {
@@ -29,8 +29,8 @@
 		"node-sass": "^7.0.1",
 		"scss": "^0.2.4",
 		"svelte-resize-observer": "^2.0.0",
-		"sveltekit-i18n": "^2.2.2",
-		"vite": "^3.0.0",
+		"sveltekit-i18n": "^2.3.0",
+		"vite": "^6.0.0",
 		"webmidi": "^3.0.21"
 	}
 }
