+++ src/routes/+layout.svelte
@@ -1,0 +1,2 @@
+<slot></slot>
+
